Make debug printing work with historic game data
With debug mode enabled, the time since start of game is displayed
for each game. The data which is returned from "games/updates" doesn't
return the start/end time for a game, so this was excepting.

Added a stopgap to get similar functionality by just logging the
timestamp of the event -- not sure if this'll actually be any use.